minor changes for pyright

+++ src/function_classes/wrappers.py
@@ -29,7 +29,7 @@ class ScaledRegression(ModifiedFunctionClass):
        ScaledXRegression or ScaledYRegression"""
     def __init__(
             self,
-            scale: int,
+            scale: float,
             inner_function_class: FunctionClass
         ):
         super(ScaledRegression, self).__init__(inner_function_class)
